pyproject.toml
[project]
name = "gaokao-vault"
version = "0.0.1"
description = "A modern full-volume Gaokao data vault: scrape, store and query comprehensive historical & latest admission data, scores, plans"
authors = [{ name = "lifefloating", email = "imshuazi@126.com" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.10,<4.0"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "scrapling[all]>=0.4.3",
    "asyncpg>=0.30",
    "typer>=0.15",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyfreeproxy>=0.4",
    "fake-useragent>=2.0",
]

[project.scripts]
gaokao-vault = "gaokao_vault.cli:app"

[project.urls]
Homepage = "https://github.com/lifefloating/gaokao-vault"
Repository = "https://github.com/lifefloating/gaokao-vault"
Documentation = "https://lifefloating.github.io/gaokao-vault/"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pre-commit>=4.5.1",
    "tox-uv>=1.33.4",
    "deptry>=0.24.0",
    "ty>=0.0.23",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.6",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.5",
    "mkdocstrings[python]>=1.0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gaokao_vault"]

[tool.ty.environment]
python = "./.venv"
python-version = "3.10"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 120
fix = true
exclude = ["skills/scrapling-official"]

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["src"]
